refactor(spelunk-core): split config.rs into config/{mod,sync_mode,project_id,paths,persist,predicates,tls}.rs
Pure move, no behaviour change. config.rs had grown to 2655 lines (~60%
tests); split by cohesion into config/mod.rs (Config/AuthTokens/IndexConfig
aggregator + load/validate/resolve_* impls) plus six leaf submodules.
secret_store.rs is untouched.

Public API is preserved byte-for-byte via pub use re-exports from mod.rs,
including apply_server_ca (new config/tls.rs, a zero-sibling-dep leaf and
the only reqwest surface in config). Two internal helpers cross a submodule
boundary and go private -> pub(in crate::config): spelunk_config_dir and
find_project_config. Tests are co-located with their code; four submodules
(project_id, paths, persist, tls) get their own #[cfg(test)] use
tempfile::TempDir since that import is file-level, not inherited via
use super::*.

Verified move-only via a sorted-multiset line diff against origin/main's
config.rs: the only removed lines are the two visibility bumps, and every
added line is a mod/use/pub-use/test-scaffold addition.
