+++ crates/spelunk-core/src/config/mod.rs
@@ -2,255 +2,33 @@ use anyhow::{Context, Result};
 use serde::{Deserialize, Serialize};
 use std::path::{Path, PathBuf};
 
+mod paths;
+mod persist;
+mod predicates;
+mod project_id;
+mod sync_mode;
+mod tls;
+
 pub mod secret_store;
 
+use paths::{find_project_config, spelunk_config_dir};
 use secret_store::{KEY_SERVER_KEY, SecretStore};
+
+pub use paths::{
+    find_project_db, find_project_dir, require_project_db, require_project_db_at, resolve_db,
+};
+pub use persist::{
+    remove_auth_tokens, remove_auth_tokens_from, remove_server_key, remove_server_key_from,
+    remove_server_key_with, save_auth_tokens, save_auth_tokens_to, save_server_key,
+    save_server_key_with, write_project_slug,
+};
+pub use predicates::{is_loopback_url, looks_like_uuid, no_server_env_set, validate_transport_url};
+pub use project_id::derive_project_id;
+pub use sync_mode::SyncMode;
+pub use tls::apply_server_ca;
 
 #[cfg(test)]
 use tempfile::TempDir;
-
-// ---------------------------------------------------------------------------
-// Sync mode (ADR-037 D1)
-// ---------------------------------------------------------------------------
-
-/// Persistent, per-project control over where memory reads/writes go and whether
-/// the CLI ever contacts the cloud (ADR-037 D1).
-///
-/// Replaces the implicit "is the server reachable" branch that previously drove
-/// backend selection. The mode is resolved once from config + environment (see
-/// [`Config::resolve_mode`]) and then gates both the capability tier probe and
-/// the memory backend selector.
-///
-/// | mode          | reads          | writes                    | cloud contact            |
-/// |---------------|----------------|---------------------------|--------------------------|
-/// | `offline`     | local          | local                     | never (even if `server_url` set) |
-/// | `local_first` | local          | local, then async background sync | best-effort              |
-/// | `cloud_first` | server (error if unreachable) | server (error if unreachable) | required |
-///
-/// `cloud_first` is the **server-authoritative** option: it is a deliberate
-/// override of ADR-005's local-as-source-of-truth invariant, and an
-/// unreachable or untrusted server is a hard error. There is no silent local
-/// fallback and no local write queue.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum SyncMode {
-    /// Local-only: never contacts the cloud, even when `server_url` is set.
-    /// The provable no-cloud guarantee for OSS testing and air-gapped use.
-    Offline,
-    /// Default when `server_url` is set: reads and writes are local; a best-effort
-    /// background sync converges the cloud replica. Offline-resilient.
-    LocalFirst,
-    /// Server-authoritative: reads and writes go to the server, and an
-    /// unreachable or untrusted server is a hard error. No silent local
-    /// fallback, no local write queue.
-    CloudFirst,
-}
-
-impl SyncMode {
-    /// Parse a mode from its serialized string form (case-insensitive).
-    ///
-    /// Accepts `offline`, `local_first`, and `cloud_first`. Returns `None` for
-    /// any other value so callers can decide how to handle an invalid override.
-    pub fn parse(s: &str) -> Option<Self> {
-        match s.trim().to_ascii_lowercase().as_str() {
-            "offline" => Some(Self::Offline),
-            "local_first" | "local-first" | "localfirst" => Some(Self::LocalFirst),
-            "cloud_first" | "cloud-first" | "cloudfirst" => Some(Self::CloudFirst),
-            _ => None,
-        }
-    }
-
-    /// String form used in config files and `SPELUNK_MODE`.
-    pub fn as_str(&self) -> &'static str {
-        match self {
-            Self::Offline => "offline",
-            Self::LocalFirst => "local_first",
-            Self::CloudFirst => "cloud_first",
-        }
-    }
-
-    /// Whether this mode permits any contact with the cloud server.
-    ///
-    /// `offline` never contacts the cloud; the other two may. Used by the
-    /// capability tier probe and the memory backend selector to honour the
-    /// kill-switch semantics of `offline`.
-    pub fn allows_cloud(&self) -> bool {
-        !matches!(self, Self::Offline)
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Project-id derivation
-// ---------------------------------------------------------------------------
-
-/// Derive a stable project identifier from `project_root`.
-///
-/// 1. Read `remote.origin.url` from the git config and normalise to
-///    `host/owner/repo`.
-/// 2. If no git repo or no origin remote, fall back to
-///    `local/<blake3-hex-of-canonical-path>`.
-pub fn derive_project_id(project_root: &Path) -> String {
-    try_derive_from_git(project_root).unwrap_or_else(|| derive_local_fallback(project_root))
-}
-
-fn try_derive_from_git(root: &Path) -> Option<String> {
-    let repo = gix::discover(root).ok()?;
-    let git_dir = repo.git_dir();
-
-    // For linked worktrees the config lives in the main .git dir, not
-    // .git/worktrees/<name>.
-    let config_path = if git_dir.parent().and_then(|p| p.file_name())
-        == Some(std::ffi::OsStr::new("worktrees"))
-    {
-        git_dir.parent()?.parent()?.join("config")
-    } else {
-        git_dir.join("config")
-    };
-
-    let content = std::fs::read_to_string(config_path).ok()?;
-    let url = extract_origin_url_from_git_config(&content)?;
-    Some(normalise_git_url(&url))
-}
-
-/// Minimal parser for git config: finds `url` under `[remote "origin"]`.
-fn extract_origin_url_from_git_config(config: &str) -> Option<String> {
-    let mut in_origin = false;
-    for line in config.lines() {
-        let trimmed = line.trim();
-        if trimmed.starts_with('[') {
-            // Section header — check if it's [remote "origin"]
-            let header = trimmed.trim_start_matches('[').trim_end_matches(']');
-            in_origin = header.trim() == r#"remote "origin""#;
-        } else if in_origin
-            && let Some(rest) = trimmed.strip_prefix("url")
-            && let Some(rest) = rest.trim_start().strip_prefix('=')
-        {
-            return Some(rest.trim().to_string());
-        }
-    }
-    None
-}
-
-/// Normalise a git remote URL to `host/owner/repo` (no scheme, no `.git`).
-///
-/// Handles `https://`, `ssh://`, and SCP-style `git@host:owner/repo.git`.
-fn normalise_git_url(url: &str) -> String {
-    let without_scheme = if let Some(pos) = url.find("://") {
-        &url[pos + 3..]
-    } else {
-        url
-    };
-    let without_user = if let Some(pos) = without_scheme.find('@') {
-        &without_scheme[pos + 1..]
-    } else {
-        without_scheme
-    };
-    // SCP colon → slash
-    let normalised = without_user.replacen(':', "/", 1);
-    let normalised = normalised.strip_suffix(".git").unwrap_or(&normalised);
-    normalised.to_lowercase()
-}
-
-fn derive_local_fallback(root: &Path) -> String {
-    let canonical = crate::utils::canonicalize(root);
-    let hash = blake3::hash(canonical.to_string_lossy().as_bytes());
-    format!("local/{}", hash.to_hex())
-}
-
-/// Returns `~/.config/spelunk/`.
-///
-/// On all platforms we use `~/.config` rather than the OS-native config dir
-/// (e.g. `~/Library/Application Support` on macOS) so that the path matches
-/// what the CLI documentation and error messages say, and so that config files
-/// work the same way across Linux and macOS.
-fn spelunk_config_dir() -> PathBuf {
-    dirs::home_dir()
-        .unwrap_or_else(|| PathBuf::from("."))
-        .join(".config")
-        .join("spelunk")
-}
-
-/// Walk up from `start` looking for `.spelunk/index.db`.
-/// Returns the first match found, or `None` if the filesystem root is reached.
-///
-/// If `start` is inside a git linked worktree, the walk begins from the main
-/// worktree root so that linked worktrees share the same index without a symlink.
-pub fn find_project_db(start: &Path) -> Option<PathBuf> {
-    let resolved = crate::utils::resolve_main_worktree_root(start);
-    let mut dir = resolved;
-    loop {
-        let candidate = dir.join(".spelunk").join("index.db");
-        if candidate.exists() {
-            return Some(candidate);
-        }
-        if !dir.pop() {
-            return None;
-        }
-    }
-}
-
-/// Walk up from `start` for a `.spelunk/` **directory** (worktree-aware).
-/// Returns the `.spelunk` dir path, or `None` if none is found before the root.
-///
-/// Keys on the directory, not `index.db`: `spelunk init --no-index` writes a
-/// `.spelunk/config.toml` with no index, and memory needs no index. Linked
-/// worktrees resolve to the main worktree's `.spelunk/` (mirrors
-/// [`find_project_db`]).
-pub fn find_project_dir(start: &Path) -> Option<PathBuf> {
-    let mut dir = crate::utils::resolve_main_worktree_root(start);
-    loop {
-        let candidate = dir.join(".spelunk");
-        if candidate.is_dir() {
-            return Some(candidate);
-        }
-        if !dir.pop() {
-            return None;
-        }
-    }
-}
-
-/// ADR-067: resolve the local project's `.spelunk/index.db` base path anchored at
-/// `start`, failing closed when `start` has no `.spelunk/` project instead of
-/// silently falling back to the global `~/.config/spelunk/` store.
-///
-/// Explicit `--db` / index-path callers bypass this (an explicit store is always
-/// honored). Memory callers apply `.with_file_name("memory.db")` to the result.
-/// `allow_global` restores the legacy global fallback and is reserved for a
-/// future `--global` flag (ADR-067 D2); no caller sets it today.
-pub fn require_project_db_at(
-    start: &Path,
-    cfg_default: &Path,
-    allow_global: bool,
-) -> Result<PathBuf> {
-    if let Some(spelunk_dir) = find_project_dir(start) {
-        return Ok(spelunk_dir.join("index.db"));
-    }
-    if allow_global {
-        return Ok(cfg_default.to_path_buf());
-    }
-    anyhow::bail!("no spelunk project here. Run 'spelunk init' first")
-}
-
-/// [`require_project_db_at`] anchored at the current working directory.
-pub fn require_project_db(cfg_default: &Path, allow_global: bool) -> Result<PathBuf> {
-    let cwd = std::env::current_dir().context("resolving current directory")?;
-    require_project_db_at(&cwd, cfg_default, allow_global)
-}
-
-/// Walk up from `start` looking for `.spelunk/config.toml` (project-level config).
-/// Stops at the filesystem root. Returns the path if found.
-fn find_project_config(start: &Path) -> Option<PathBuf> {
-    let mut dir = start.to_path_buf();
-    loop {
-        let candidate = dir.join(".spelunk").join("config.toml");
-        if candidate.exists() {
-            return Some(candidate);
-        }
-        if !dir.pop() {
-            return None;
-        }
-    }
-}
 
 /// Serde default helper: `true`.
 fn default_true() -> bool {
@@ -311,21 +89,6 @@ struct ProjectConfig {
     server_ca: Option<String>,
     /// `[index]` table: per-field override of the built-in file filter.
     index: Option<ProjectIndexConfig>,
-}
-
-/// Resolve the database path.
-///
-/// Priority: explicit `--db` arg > project DB (walk up from CWD) > `cfg_default`.
-pub fn resolve_db(explicit: Option<&Path>, cfg_default: &Path) -> PathBuf {
-    if let Some(p) = explicit {
-        return p.to_path_buf();
-    }
-    if let Ok(cwd) = std::env::current_dir()
-        && let Some(p) = find_project_db(&cwd)
-    {
-        return p;
-    }
-    cfg_default.to_path_buf()
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -763,304 +526,6 @@ impl Config {
     }
 }
 
-/// Persist the CLI bearer credential (`server_key`) into the OS secret store.
-///
-/// This is the token `spelunk login` persists. It is stored in the OS keychain
-/// (macOS Keychain / Linux Secret Service / Windows Credential Manager) by
-/// default, falling back to an owner-only file when no keychain backend is
-/// available (CI / headless). The credential is **never** written to
-/// `config.toml`.
-///
-/// The store is credential-format-agnostic, so the same call path serves a
-/// future WorkOS-token migration.
-pub fn save_server_key(key: &str) -> Result<()> {
-    let store = secret_store::default_store(&spelunk_config_dir())?;
-    save_server_key_with(key, store.as_ref())
-}
-
-/// Same as [`save_server_key`] but with an injected [`SecretStore`] (tests).
-pub fn save_server_key_with(key: &str, store: &dyn SecretStore) -> Result<()> {
-    store
-        .set(KEY_SERVER_KEY, key)
-        .context("persisting server_key to the secret store")
-}
-
-/// Remove the CLI bearer credential everywhere it might live.
-///
-/// What `spelunk logout` clears: the secret-store entry **and** any legacy
-/// plaintext `server_key` left in `~/.config/spelunk/config.toml`. No-op for any
-/// location where the credential is absent.
-pub fn remove_server_key() -> Result<()> {
-    let store = secret_store::default_store(&spelunk_config_dir())?;
-    remove_server_key_with(store.as_ref(), &spelunk_config_dir().join("config.toml"))
-}
-
-/// Same as [`remove_server_key`] but with an injected [`SecretStore`] and an
-/// explicit legacy config path (tests).
-pub fn remove_server_key_with(store: &dyn SecretStore, config_path: &Path) -> Result<()> {
-    store
-        .delete(KEY_SERVER_KEY)
-        .context("removing server_key from the secret store")?;
-    // Belt-and-braces: also strip any legacy plaintext key still in the file.
-    remove_server_key_from(config_path)
-}
-
-/// Same as [`remove_server_key`] but operates on an explicit path (useful in tests).
-pub fn remove_server_key_from(config_path: &Path) -> Result<()> {
-    if !config_path.exists() {
-        return Ok(());
-    }
-    let existing = std::fs::read_to_string(config_path)
-        .with_context(|| format!("reading {}", config_path.display()))?;
-
-    let updated: String = existing
-        .lines()
-        .filter(|line| {
-            let trimmed = line.trim();
-            // Keep every line that is NOT a `server_key = …` assignment.
-            // strip_prefix avoids a raw byte-index slice and handles the
-            // "server_key" prefix unambiguously.
-            let after_key = trimmed.strip_prefix("server_key");
-            !matches!(after_key, Some(rest) if rest.trim_start().starts_with('='))
-        })
-        .map(|line| format!("{line}\n"))
-        .collect();
-
-    std::fs::write(config_path, updated)
-        .with_context(|| format!("writing {}", config_path.display()))?;
-    Ok(())
-}
-
-// ───────────────────────────────────────────────────────────────────────────
-// `[auth]` table persistence (WorkOS device-flow tokens, ADR-045)
-// ───────────────────────────────────────────────────────────────────────────
-
-/// Persist WorkOS tokens to the `[auth]` table of `~/.config/spelunk/config.toml`.
-///
-/// Replaces any existing `[auth]` table; all other top-level keys and tables
-/// are preserved. The file is written with `0600` permissions so the refresh
-/// token is not world-readable.
-pub fn save_auth_tokens(tokens: &AuthTokens) -> Result<()> {
-    save_auth_tokens_to(tokens, &spelunk_config_dir().join("config.toml"))
-}
-
-/// Same as [`save_auth_tokens`] but writes to an explicit path (useful in tests).
-pub fn save_auth_tokens_to(tokens: &AuthTokens, config_path: &Path) -> Result<()> {
-    if let Some(parent) = config_path.parent() {
-        std::fs::create_dir_all(parent)
-            .with_context(|| format!("creating config dir {}", parent.display()))?;
-    }
-
-    let mut doc = read_config_table(config_path)?;
-    let auth_value = toml::Value::try_from(tokens).context("serialising auth tokens")?;
-    doc.insert("auth".to_string(), auth_value);
-
-    let serialised = toml::to_string_pretty(&doc).context("serialising config.toml")?;
-    write_config_secure(config_path, &serialised)
-}
-
-/// Remove the `[auth]` table from `~/.config/spelunk/config.toml`.
-///
-/// What `spelunk logout` clears (alongside the legacy `server_key`). No-op if
-/// the file or the table is absent. Other keys are preserved.
-pub fn remove_auth_tokens() -> Result<()> {
-    remove_auth_tokens_from(&spelunk_config_dir().join("config.toml"))
-}
-
-/// Same as [`remove_auth_tokens`] but operates on an explicit path (tests).
-pub fn remove_auth_tokens_from(config_path: &Path) -> Result<()> {
-    if !config_path.exists() {
-        return Ok(());
-    }
-    let mut doc = read_config_table(config_path)?;
-    if doc.remove("auth").is_none() {
-        return Ok(());
-    }
-    let serialised = toml::to_string_pretty(&doc).context("serialising config.toml")?;
-    write_config_secure(config_path, &serialised)
-}
-
-/// Write `slug` as `project_id` to a project-level `.spelunk/config.toml`,
-/// creating the file (and parent dir) if absent. Other keys are preserved.
-///
-/// Returns `(effective_slug, wrote)`: if the file already sets `project_id`,
-/// it is left untouched — `wrote` is `false` and the existing value is returned
-/// (no retroactive rename). This is a committed, shared file, so it is written
-/// with normal permissions (unlike the secret-bearing personal config).
-pub fn write_project_slug(config_path: &Path, slug: &str) -> Result<(String, bool)> {
-    let mut doc = read_config_table(config_path)?;
-    if let Some(existing) = doc.get("project_id").and_then(|v| v.as_str()) {
-        return Ok((existing.to_string(), false));
-    }
-    if let Some(parent) = config_path.parent() {
-        std::fs::create_dir_all(parent)
-            .with_context(|| format!("creating config dir {}", parent.display()))?;
-    }
-    doc.insert(
-        "project_id".to_string(),
-        toml::Value::String(slug.to_string()),
-    );
-    let serialised = toml::to_string_pretty(&doc).context("serialising config.toml")?;
-    std::fs::write(config_path, serialised)
-        .with_context(|| format!("writing {}", config_path.display()))?;
-    Ok((slug.to_string(), true))
-}
-
-/// Parse the config file into a `toml::Table`, returning an empty table when the
-/// file does not exist.
-fn read_config_table(config_path: &Path) -> Result<toml::Table> {
-    if !config_path.exists() {
-        return Ok(toml::Table::new());
-    }
-    let raw = std::fs::read_to_string(config_path)
-        .with_context(|| format!("reading {}", config_path.display()))?;
-    raw.parse::<toml::Table>()
-        .with_context(|| format!("parsing {}", config_path.display()))
-}
-
-/// Write `contents` to `config_path` and tighten permissions to `0600` on Unix
-/// so secrets in the file are owner-only.
-fn write_config_secure(config_path: &Path, contents: &str) -> Result<()> {
-    std::fs::write(config_path, contents)
-        .with_context(|| format!("writing {}", config_path.display()))?;
-    set_owner_only_permissions(config_path)?;
-    Ok(())
-}
-
-/// Set `0600` permissions on Unix; a no-op on other platforms.
-#[cfg(unix)]
-fn set_owner_only_permissions(path: &Path) -> Result<()> {
-    use std::os::unix::fs::PermissionsExt;
-    let perms = std::fs::Permissions::from_mode(0o600);
-    std::fs::set_permissions(path, perms)
-        .with_context(|| format!("setting 0600 permissions on {}", path.display()))?;
-    Ok(())
-}
-
-#[cfg(not(unix))]
-fn set_owner_only_permissions(_path: &Path) -> Result<()> {
-    Ok(())
-}
-
-/// Returns `true` when `SPELUNK_NO_SERVER` is set to a truthy value.
-///
-/// This is the hard offline kill-switch shared by [`Config::resolve_mode`] and
-/// the CLI capability probe; both must agree on what "no server" means.
-pub fn no_server_env_set() -> bool {
-    matches!(
-        std::env::var("SPELUNK_NO_SERVER").as_deref(),
-        Ok("1") | Ok("true") | Ok("yes")
-    )
-}
-
-/// Return `true` if `s` parses as a canonical UUID (any version).
-///
-/// Used by the cloud-api slug→UUID resolution path (ADR-005): a `project_id`
-/// that is already a UUID is used directly against `/v1/projects/{uuid}/…`,
-/// while a non-UUID value is treated as a human slug and resolved via
-/// `GET /v1/projects`.
-pub fn looks_like_uuid(s: &str) -> bool {
-    uuid::Uuid::parse_str(s).is_ok()
-}
-
-/// Return `true` if `url` targets a loopback address (`127.x.x.x`, `localhost`, `::1`).
-///
-/// This is a lightweight string check — no DNS resolution.
-pub fn is_loopback_url(url: &str) -> bool {
-    // Strip scheme and authority prefix up to the host.
-    let host_part = url
-        .trim_start_matches("http://")
-        .trim_start_matches("https://");
-
-    // Extract the host (before any path or port).
-    let host = if let Some(idx) = host_part.find('/') {
-        &host_part[..idx]
-    } else {
-        host_part
-    };
-    // Drop port if present (handle IPv6 bracketed form too).
-    let host = if host.starts_with('[') {
-        // IPv6: [::1]:port or [::1]
-        host.trim_start_matches('[')
-            .split(']')
-            .next()
-            .unwrap_or(host)
-    } else {
-        host.split(':').next().unwrap_or(host)
-    };
-
-    matches!(host, "localhost" | "127.0.0.1" | "::1") || host.starts_with("127.")
-}
-
-/// Validate that `url` is an acceptable transport for sending a bearer token /
-/// talking to a spelunk-server: either `https://` (any host), or `http://` to a
-/// loopback host (`127.0.0.1`, `::1`, `localhost`).
-///
-/// A non-loopback `http://` URL is invalid config — plaintext HTTP outside the
-/// loopback interface would send the bearer token (and query content) in the
-/// clear. There is no opt-out env var: the fix is always "use https, or
-/// loopback".
-///
-/// Like [`is_loopback_url`], this is a lightweight string check on the literal
-/// host — there is no DNS resolution. A `/etc/hosts` alias or other custom DNS
-/// entry that resolves to a loopback address but isn't spelled `127.x.x.x`,
-/// `::1`, or `localhost` is **not** recognised as loopback and is rejected;
-/// this is intentional (fail closed, not open) and the known limitation of a
-/// string-based check.
-///
-/// Returns `Ok(())` for a valid URL, or a one-line `Err` naming the fix.
-pub fn validate_transport_url(url: &str) -> Result<(), String> {
-    if url.starts_with("https://") {
-        return Ok(());
-    }
-    if url.starts_with("http://") {
-        if is_loopback_url(url) {
-            return Ok(());
-        }
-        return Err(format!(
-            "invalid server URL {url:?}: plaintext http:// is only allowed to a loopback \
-             address (127.0.0.1/::1/localhost); use https:// for any other host"
-        ));
-    }
-    Err(format!(
-        "invalid server URL {url:?}: expected an http:// or https:// URL"
-    ))
-}
-
-/// Apply the configured custom CA bundle to a reqwest client builder.
-///
-/// `ca_path` is the resolved [`Config::server_ca`] (env `SPELUNK_SERVER_CA`
-/// precedence is already applied at load time). Adds every certificate in the
-/// PEM bundle as a trust anchor **on top of** the built-in roots; certificate
-/// verification stays on. A `None` path is a no-op, so every
-/// team-server client site can route through this unconditionally.
-pub fn apply_server_ca(
-    builder: reqwest::ClientBuilder,
-    ca_path: Option<&Path>,
-) -> Result<reqwest::ClientBuilder> {
-    let Some(path) = ca_path else {
-        return Ok(builder);
-    };
-    let pem = std::fs::read(path)
-        .with_context(|| format!("reading SPELUNK_SERVER_CA bundle at {}", path.display()))?;
-    let certs = reqwest::Certificate::from_pem_bundle(&pem)
-        .with_context(|| format!("parsing PEM CA bundle at {}", path.display()))?;
-    // `from_pem_bundle` yields an empty vec for a file with no PEM blocks rather
-    // than erroring — surface that as a config error, else a wrong path would
-    // silently add no trust anchor and fail TLS with a confusing message.
-    if certs.is_empty() {
-        anyhow::bail!(
-            "no PEM certificates found in CA bundle at {}",
-            path.display()
-        );
-    }
-    let mut builder = builder;
-    for cert in certs {
-        builder = builder.add_root_certificate(cert);
-    }
-    Ok(builder)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1072,58 +537,6 @@ mod tests {
         Config::load_with_store(Some(path), &MemoryStore::default())
     }
 
-    #[test]
-    fn write_project_slug_creates_file_and_reports_written() {
-        let tmp = TempDir::new().unwrap();
-        let cfg = tmp.path().join(".spelunk").join("config.toml");
-
-        let (slug, wrote) = write_project_slug(&cfg, "github.com/acme/app").unwrap();
-        assert_eq!(slug, "github.com/acme/app");
-        assert!(wrote);
-
-        let raw = std::fs::read_to_string(&cfg).unwrap();
-        assert!(
-            raw.contains("project_id = \"github.com/acme/app\""),
-            "{raw}"
-        );
-    }
-
-    #[test]
-    fn write_project_slug_preserves_existing_and_does_not_rename() {
-        let tmp = TempDir::new().unwrap();
-        let cfg = tmp.path().join("config.toml");
-        std::fs::write(
-            &cfg,
-            "server_url = \"http://team:7777\"\nproject_id = \"old/slug\"\n",
-        )
-        .unwrap();
-
-        let (slug, wrote) = write_project_slug(&cfg, "local/deadbeef").unwrap();
-        assert_eq!(slug, "old/slug");
-        assert!(!wrote);
-
-        let raw = std::fs::read_to_string(&cfg).unwrap();
-        assert!(raw.contains("project_id = \"old/slug\""), "{raw}");
-        assert!(raw.contains("server_url"), "other keys preserved: {raw}");
-    }
-
-    #[test]
-    fn write_project_slug_adds_key_preserving_other_keys() {
-        let tmp = TempDir::new().unwrap();
-        let cfg = tmp.path().join("config.toml");
-        std::fs::write(&cfg, "server_url = \"http://team:7777\"\n").unwrap();
-
-        let (_, wrote) = write_project_slug(&cfg, "github.com/acme/app").unwrap();
-        assert!(wrote);
-
-        let raw = std::fs::read_to_string(&cfg).unwrap();
-        assert!(
-            raw.contains("project_id = \"github.com/acme/app\""),
-            "{raw}"
-        );
-        assert!(raw.contains("server_url"), "existing key preserved: {raw}");
-    }
-
     /// Unset all spelunk-related env vars to prevent cross-test contamination.
     fn clear_spelunk_env() {
         unsafe {
@@ -1133,45 +546,6 @@ mod tests {
             std::env::remove_var("SPELUNK_MODE");
             std::env::remove_var("SPELUNK_NO_SERVER");
         }
-    }
-
-    // ── SyncMode parse / as_str (ADR-037 D1) ─────────────────────────────────
-
-    #[test]
-    fn sync_mode_parse_accepts_canonical_and_variant_forms() {
-        assert_eq!(SyncMode::parse("offline"), Some(SyncMode::Offline));
-        assert_eq!(SyncMode::parse("LOCAL_FIRST"), Some(SyncMode::LocalFirst));
-        assert_eq!(SyncMode::parse("local-first"), Some(SyncMode::LocalFirst));
-        assert_eq!(SyncMode::parse("cloud_first"), Some(SyncMode::CloudFirst));
-        assert_eq!(SyncMode::parse(" cloudfirst "), Some(SyncMode::CloudFirst));
-        assert_eq!(SyncMode::parse("bogus"), None);
-    }
-
-    #[test]
-    fn sync_mode_as_str_round_trips() {
-        for m in [
-            SyncMode::Offline,
-            SyncMode::LocalFirst,
-            SyncMode::CloudFirst,
-        ] {
-            assert_eq!(SyncMode::parse(m.as_str()), Some(m));
-        }
-    }
-
-    #[test]
-    fn sync_mode_allows_cloud() {
-        assert!(!SyncMode::Offline.allows_cloud());
-        assert!(SyncMode::LocalFirst.allows_cloud());
-        assert!(SyncMode::CloudFirst.allows_cloud());
-    }
-
-    #[test]
-    fn sync_mode_serde_snake_case() {
-        // Serialised form must be snake_case so config.toml / wire stay stable.
-        let json = serde_json::to_string(&SyncMode::LocalFirst).unwrap();
-        assert_eq!(json, "\"local_first\"");
-        let parsed: SyncMode = serde_json::from_str("\"cloud_first\"").unwrap();
-        assert_eq!(parsed, SyncMode::CloudFirst);
     }
 
     // ── resolve_mode defaults (ADR-037 D1) ───────────────────────────────────
@@ -1474,193 +848,6 @@ memory_server_key = "old-token"
         assert!(without_id.validate().is_err());
     }
 
-    // ── is_loopback_url ──────────────────────────────────────────────────────
-
-    #[test]
-    fn is_loopback_url_recognises_127_0_0_1() {
-        assert!(is_loopback_url("http://127.0.0.1:7777"));
-        assert!(is_loopback_url("http://127.0.0.1:7777/"));
-        assert!(is_loopback_url("http://127.0.0.1"));
-    }
-
-    #[test]
-    fn is_loopback_url_recognises_localhost() {
-        assert!(is_loopback_url("http://localhost:7777"));
-        assert!(is_loopback_url("http://localhost"));
-    }
-
-    #[test]
-    fn is_loopback_url_recognises_ipv6_loopback() {
-        assert!(is_loopback_url("http://[::1]:7777"));
-        assert!(is_loopback_url("http://[::1]"));
-    }
-
-    #[test]
-    fn is_loopback_url_recognises_127_subnet() {
-        assert!(is_loopback_url("http://127.1.2.3:7777"));
-    }
-
-    #[test]
-    fn is_loopback_url_rejects_non_loopback() {
-        assert!(!is_loopback_url("http://spelunk.internal:7777"));
-        assert!(!is_loopback_url("http://192.168.1.100:7777"));
-        assert!(!is_loopback_url("https://example.com"));
-        assert!(!is_loopback_url("http://10.0.0.1"));
-    }
-
-    #[test]
-    fn is_loopback_url_rejects_address_with_127_in_path() {
-        // Should NOT match just because "127" appears somewhere
-        assert!(!is_loopback_url("http://example.com/proxy/127.0.0.1"));
-    }
-
-    // ── validate_transport_url (loopback-only plaintext http) ──────────────────
-
-    #[test]
-    fn validate_transport_url_rejects_non_loopback_http() {
-        let err = validate_transport_url("http://team-server:7777")
-            .expect_err("non-loopback http:// must be rejected");
-        assert!(err.contains("http://team-server:7777"));
-        assert!(err.contains("https"));
-        assert!(err.contains("loopback"));
-    }
-
-    #[test]
-    fn validate_transport_url_rejects_non_loopback_ip_http() {
-        assert!(validate_transport_url("http://192.168.1.100:7777").is_err());
-        assert!(validate_transport_url("http://10.0.0.1:7777").is_err());
-    }
-
-    #[test]
-    fn validate_transport_url_accepts_loopback_http() {
-        assert!(validate_transport_url("http://127.0.0.1:7777").is_ok());
-        assert!(validate_transport_url("http://localhost:7777").is_ok());
-        assert!(validate_transport_url("http://[::1]:7777").is_ok());
-        assert!(validate_transport_url("http://127.5.6.7:7777").is_ok());
-    }
-
-    #[test]
-    fn validate_transport_url_accepts_any_https() {
-        assert!(validate_transport_url("https://team-server:7777").is_ok());
-        assert!(validate_transport_url("https://example.com").is_ok());
-        assert!(validate_transport_url("https://127.0.0.1:7777").is_ok());
-    }
-
-    #[test]
-    fn validate_transport_url_rejects_unknown_scheme() {
-        let err = validate_transport_url("ftp://team-server:7777").unwrap_err();
-        assert!(err.contains("http"));
-    }
-
-    /// IPv6 non-loopback addresses must be rejected too, not just the IPv4 case
-    /// — the check is symmetric across address families.
-    #[test]
-    fn validate_transport_url_rejects_non_loopback_ipv6_http() {
-        assert!(validate_transport_url("http://[2001:db8::1]:7777").is_err());
-        assert!(validate_transport_url("http://[fe80::1]:7777").is_err());
-    }
-
-    /// Known limitation, asserted so it can't silently regress into a security
-    /// hole: this is a *string* check with no DNS resolution. A hostname alias
-    /// (e.g. an `/etc/hosts` entry) that an OS resolver would send to
-    /// 127.0.0.1 is NOT recognised as loopback here and is correctly rejected
-    /// (fail closed) rather than accepted on the assumption it "means"
-    /// loopback. If a caller ever needs alias support, that requires an
-    /// explicit, reviewed allow-list — not a silent DNS lookup at validation
-    /// time (which would also make validation do network I/O).
-    #[test]
-    fn validate_transport_url_rejects_hostname_alias_even_if_it_would_resolve_to_loopback() {
-        let err = validate_transport_url("http://my-loopback-alias:7777")
-            .expect_err("a non-literal loopback hostname must be rejected, not DNS-resolved");
-        assert!(err.contains("loopback"));
-    }
-
-    /// `localhost` with an explicit port-less bare form and a trailing slash
-    /// path both still resolve to the same host extraction as the bracketed
-    /// IPv6 case — pin the unbracketed `::1` (no port) form too, since the
-    /// bracket-stripping logic in `is_loopback_url` is easy to regress.
-    #[test]
-    fn validate_transport_url_accepts_bare_ipv6_loopback_no_port() {
-        assert!(validate_transport_url("http://[::1]/").is_ok());
-    }
-
-    // ── looks_like_uuid (ADR-005) ────────────────────────────────────────────
-
-    #[test]
-    fn looks_like_uuid_accepts_canonical_uuids() {
-        assert!(looks_like_uuid("018f4e2a-1234-7abc-8def-000000000001"));
-        assert!(looks_like_uuid("00000000-0000-0000-0000-000000000000"));
-        // uppercase hex is valid
-        assert!(looks_like_uuid("018F4E2A-1234-7ABC-8DEF-000000000001"));
-    }
-
-    #[test]
-    fn looks_like_uuid_rejects_slugs() {
-        assert!(!looks_like_uuid("spelunk"));
-        assert!(!looks_like_uuid("acme/my-app"));
-        assert!(!looks_like_uuid("local/9f2a8b3c4d5e6f70"));
-        assert!(!looks_like_uuid(""));
-        // a UUID missing a section is not a UUID
-        assert!(!looks_like_uuid("018f4e2a-1234-7abc-8def"));
-    }
-
-    // ── normalise_git_url ────────────────────────────────────────────────────
-
-    #[test]
-    fn normalise_https_url() {
-        assert_eq!(
-            normalise_git_url("https://github.com/owner/repo.git"),
-            "github.com/owner/repo"
-        );
-    }
-
-    #[test]
-    fn normalise_scp_url() {
-        assert_eq!(
-            normalise_git_url("git@github.com:owner/repo.git"),
-            "github.com/owner/repo"
-        );
-    }
-
-    #[test]
-    fn normalise_ssh_url() {
-        assert_eq!(
-            normalise_git_url("ssh://git@github.com/owner/repo"),
-            "github.com/owner/repo"
-        );
-    }
-
-    // ── derive_project_id: no git repo → local/ fallback ─────────────────────
-
-    #[test]
-    fn derive_project_id_non_git_dir_returns_local_prefix() {
-        let tmp = TempDir::new().unwrap();
-        let id = derive_project_id(tmp.path());
-        assert!(id.starts_with("local/"), "expected local/ prefix, got {id}");
-        // blake3 hex is 64 chars
-        assert_eq!(id.len(), "local/".len() + 64);
-    }
-
-    // ── derive_project_id: git repo with origin ───────────────────────────────
-
-    #[test]
-    fn derive_project_id_git_repo_with_origin() {
-        let tmp = TempDir::new().unwrap();
-        let repo_dir = tmp.path().join("repo");
-        std::fs::create_dir_all(repo_dir.join(".git")).unwrap();
-        std::fs::write(
-            repo_dir.join(".git").join("config"),
-            "[core]\n\trepositoryformatversion = 0\n[remote \"origin\"]\n\turl = https://github.com/spelunk-cloud/spelunk.git\n\tfetch = +refs/heads/*:refs/remotes/origin/*\n",
-        )
-        .unwrap();
-        // derive_project_id falls back to local/ when gix::discover fails on
-        // a minimal fake repo, but the git-config parser should find the URL.
-        // We test the git-config parser directly instead:
-        let config = std::fs::read_to_string(repo_dir.join(".git").join("config")).unwrap();
-        let url = extract_origin_url_from_git_config(&config).unwrap();
-        assert_eq!(normalise_git_url(&url), "github.com/spelunk-cloud/spelunk");
-    }
-
     // ── resolve_project_id ───────────────────────────────────────────────────
 
     #[test]
@@ -1956,14 +1143,6 @@ project_id = "team/proj"
         assert!(cfg.auth.is_none());
         // Legacy key still present and now resolves as the bearer fallback.
         assert_eq!(cfg.server_key.as_deref(), Some("sk-legacy"));
-    }
-
-    /// `remove_auth_tokens_from` is a no-op when the file is missing.
-    #[test]
-    fn remove_auth_tokens_no_op_when_file_missing() {
-        let tmp = TempDir::new().unwrap();
-        let path = tmp.path().join("config.toml");
-        remove_auth_tokens_from(&path).unwrap();
     }
 
     /// On Unix, the config file is written `0600` after persisting tokens.
@@ -2437,192 +1616,6 @@ project_id = "team/new"
 
         let missing = tmp.path().join("does_not_exist.toml");
         assert!(!Config::llm_model_configured(Some(&missing)));
-    }
-
-    // ── find_project_dir / require_project_db_at (ADR-067) ───────────────────
-
-    /// Exact fail-closed error text (ADR-067; em dash restructured out per the
-    /// no-em-dash house rule for user-facing copy).
-    const NO_PROJECT_ERR: &str = "no spelunk project here. Run 'spelunk init' first";
-
-    #[test]
-    fn find_project_dir_finds_dot_spelunk_at_start() {
-        let tmp = TempDir::new().unwrap();
-        std::fs::create_dir_all(tmp.path().join(".spelunk")).unwrap();
-        assert_eq!(
-            find_project_dir(tmp.path()),
-            Some(tmp.path().join(".spelunk"))
-        );
-    }
-
-    #[test]
-    fn find_project_dir_walks_up_from_subdir() {
-        let tmp = TempDir::new().unwrap();
-        std::fs::create_dir_all(tmp.path().join(".spelunk")).unwrap();
-        let sub = tmp.path().join("a").join("b");
-        std::fs::create_dir_all(&sub).unwrap();
-        assert_eq!(find_project_dir(&sub), Some(tmp.path().join(".spelunk")));
-    }
-
-    #[test]
-    fn find_project_dir_none_when_absent() {
-        let tmp = TempDir::new().unwrap();
-        assert_eq!(find_project_dir(tmp.path()), None);
-    }
-
-    #[test]
-    fn find_project_dir_ignores_dot_spelunk_file() {
-        // A regular file named `.spelunk` is not a project dir.
-        let tmp = TempDir::new().unwrap();
-        std::fs::write(tmp.path().join(".spelunk"), "not a dir").unwrap();
-        assert_eq!(find_project_dir(tmp.path()), None);
-    }
-
-    #[test]
-    fn require_project_db_at_returns_scoped_index_db_with_project() {
-        let tmp = TempDir::new().unwrap();
-        std::fs::create_dir_all(tmp.path().join(".spelunk")).unwrap();
-        let global = PathBuf::from("/nonexistent/global/index.db");
-        let got = require_project_db_at(tmp.path(), &global, false).unwrap();
-        assert_eq!(got, tmp.path().join(".spelunk").join("index.db"));
-    }
-
-    #[test]
-    fn require_project_db_at_walks_up_to_project() {
-        let tmp = TempDir::new().unwrap();
-        std::fs::create_dir_all(tmp.path().join(".spelunk")).unwrap();
-        let sub = tmp.path().join("crates").join("x");
-        std::fs::create_dir_all(&sub).unwrap();
-        let global = PathBuf::from("/nonexistent/global/index.db");
-        let got = require_project_db_at(&sub, &global, false).unwrap();
-        assert_eq!(got, tmp.path().join(".spelunk").join("index.db"));
-    }
-
-    #[test]
-    fn require_project_db_at_errors_without_project() {
-        let tmp = TempDir::new().unwrap();
-        let global = PathBuf::from("/nonexistent/global/index.db");
-        let err = require_project_db_at(tmp.path(), &global, false).unwrap_err();
-        assert_eq!(err.to_string(), NO_PROJECT_ERR);
-    }
-
-    #[test]
-    fn require_project_db_at_allow_global_returns_default_when_no_project() {
-        // The reserved --global opt-in path (ADR-067 D2) restores the legacy
-        // global fallback instead of failing closed.
-        let tmp = TempDir::new().unwrap();
-        let global = tmp.path().join("global-index.db");
-        let got = require_project_db_at(tmp.path(), &global, true).unwrap();
-        assert_eq!(got, global);
-    }
-
-    /// Build a linked-worktree fixture: `<main>/.spelunk/index.db` exists and
-    /// `<wt>` is a linked worktree with NO local `.spelunk/`. Returns
-    /// `(main_root, wt_root, index_db)`. Mirrors the fixture in `utils::tests`.
-    fn linked_worktree_fixture(tmp: &TempDir) -> (PathBuf, PathBuf, PathBuf) {
-        let main_root = tmp.path().join("main");
-        let wt_root = tmp.path().join("feat-branch");
-        std::fs::create_dir_all(&main_root).unwrap();
-        std::fs::create_dir_all(&wt_root).unwrap();
-
-        // Main worktree owns the shared index.
-        std::fs::create_dir_all(main_root.join(".spelunk")).unwrap();
-        let index_db = main_root.join(".spelunk").join("index.db");
-        std::fs::write(&index_db, b"").unwrap();
-
-        // wt_root/.git is a file pointing at <main>/.git/worktrees/<name>.
-        let gitdir_entry = main_root.join(".git").join("worktrees").join("feat-branch");
-        std::fs::create_dir_all(&gitdir_entry).unwrap();
-        std::fs::write(
-            wt_root.join(".git"),
-            format!("gitdir: {}\n", gitdir_entry.display()),
-        )
-        .unwrap();
-
-        (main_root, wt_root, index_db)
-    }
-
-    #[test]
-    fn find_project_db_resolves_worktree_to_main_index() {
-        // A linked worktree with no local `.spelunk/` resolves reads to the
-        // main worktree's shared index, with no setup step.
-        let tmp = TempDir::new().unwrap();
-        let (_main_root, wt_root, index_db) = linked_worktree_fixture(&tmp);
-        assert!(
-            !wt_root.join(".spelunk").exists(),
-            "worktree must have no local .spelunk/"
-        );
-        assert_eq!(find_project_db(&wt_root), Some(index_db));
-    }
-
-    #[test]
-    fn require_project_db_at_resolves_worktree_to_main_index() {
-        let tmp = TempDir::new().unwrap();
-        let (main_root, wt_root, _index_db) = linked_worktree_fixture(&tmp);
-        let global = PathBuf::from("/nonexistent/global/index.db");
-        let got = require_project_db_at(&wt_root, &global, false).unwrap();
-        assert_eq!(got, main_root.join(".spelunk").join("index.db"));
-    }
-
-    // ── Custom CA trust (SPELUNK_SERVER_CA / config `server_ca`) ─────────────
-
-    /// A throwaway self-signed CA used only to prove the PEM is parsed and
-    /// accepted as a trust anchor. Not trusted by anything real.
-    const TEST_CA_PEM: &[u8] = b"-----BEGIN CERTIFICATE-----\n\
-MIIDFTCCAf2gAwIBAgIUdz5ZLoL+3T+MwWN0dJjElxlwsRwwDQYJKoZIhvcNAQEL\n\
-BQAwGjEYMBYGA1UEAwwPc3BlbHVuay10ZXN0LWNhMB4XDTI2MDcxMzE3MjkyMFoX\n\
-DTM2MDcxMDE3MjkyMFowGjEYMBYGA1UEAwwPc3BlbHVuay10ZXN0LWNhMIIBIjAN\n\
-BgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAz/BAzTJJbgWUWnUqV0qFJHT+TIDT\n\
-WQJbIRVBb9MezLblAGun2RG22U47jubOKoSa4DrenrEJIafd74IR9aLUdcRp6lyN\n\
-WsuzY6P26ntZ1epHUjYeBgqpu71v3FK2pBvQ9PP//AhQN7apE6V4UocKd7OxbSk7\n\
-g1bZSYSXoFQtSZzV9KCWNpuqUMNdaMIoy1EYY86t55jeDdpFRkiO3W5jZ6M37ekg\n\
-mDq5wIOC1QHziDLWFkpBbuOxsN/admbwbsDH5301H3P25RBY12Guqsz4/lgsEuN9\n\
-L+RJfs/Vdmen5wKhbPDkr8EYx7hLF0T2ZKOf0TrJojrqHkO5n4+7ESeaUwIDAQAB\n\
-o1MwUTAdBgNVHQ4EFgQUJsLeVcwx4exuV//vdoLfqb5H3ZQwHwYDVR0jBBgwFoAU\n\
-JsLeVcwx4exuV//vdoLfqb5H3ZQwDwYDVR0TAQH/BAUwAwEB/zANBgkqhkiG9w0B\n\
-AQsFAAOCAQEAT5lW043iyZlbYM0372z/Ec8Z3VYDZ3bvryKN+6kGYuZJJnCep2c/\n\
-QX2iPx+HRWx0rz+QcnNrOdetr2KAac6ODxU2LVzjehac5wUVWm6uICzojjy84Ztn\n\
-1t5Ori6kvPSbOxJbznQuC7FILxpZswOBh6qfOHNgKeGVK4OkG2069YiFI+kwMdkI\n\
-d9qQF0w9nfELOC5M+ZxwP4vE/QkXLG57ZrOvKl2V4pthKSBv3LBAnh/C7X7/KC+f\n\
-iwNpumIaYRGylEbxW2WVv9YsWDmTBFqEkgrmx1QPJr3FtA6eeWmZ+EJIr3ImOv/d\n\
-CPBfHwWj/FUeFj+csF5QpOj+u/D1F1Kh5w==\n\
------END CERTIFICATE-----\n";
-
-    #[test]
-    fn apply_server_ca_none_is_noop() {
-        // No path → builder unchanged and still buildable.
-        let client = apply_server_ca(reqwest::Client::builder(), None)
-            .unwrap()
-            .build();
-        assert!(client.is_ok());
-    }
-
-    #[test]
-    fn apply_server_ca_adds_valid_bundle() {
-        let tmp = TempDir::new().unwrap();
-        let ca = tmp.path().join("ca.pem");
-        std::fs::write(&ca, TEST_CA_PEM).unwrap();
-        // A valid PEM bundle must parse and be accepted as a trust anchor; the
-        // client (verification still on) builds successfully.
-        let client = apply_server_ca(reqwest::Client::builder(), Some(&ca))
-            .expect("valid CA bundle should be accepted")
-            .build();
-        assert!(client.is_ok());
-    }
-
-    #[test]
-    fn apply_server_ca_missing_file_errors() {
-        let missing = Path::new("/nonexistent/spelunk-server-ca.pem");
-        let err = apply_server_ca(reqwest::Client::builder(), Some(missing)).unwrap_err();
-        assert!(err.to_string().contains("SPELUNK_SERVER_CA"), "got: {err}");
-    }
-
-    #[test]
-    fn apply_server_ca_malformed_pem_errors() {
-        let tmp = TempDir::new().unwrap();
-        let ca = tmp.path().join("bad.pem");
-        std::fs::write(&ca, b"not a certificate").unwrap();
-        assert!(apply_server_ca(reqwest::Client::builder(), Some(&ca)).is_err());
     }
 
     #[test]
